First implementation for vehicle mooving It is not working because the section or the lane that is paired to the vehicle is null.

+++ Cassini/src/org/insa/core/roadnetwork/Section.java
@@ -114,7 +114,7 @@ public class Section {
         this.sourceNode = sourceNode;
     }
     
-    public double getLength() {
+    public float getLength() {
         return length;
     }
